+++ src/stages/stage6/background_anim.h
@@ -1,0 +1,18 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
+ */
+
+#ifndef IGUARD_stages_stage6_background_anim_h
+#define IGUARD_stages_stage6_background_anim_h
+
+#include "taisei.h"
+
+#include "coroutine.h"
+
+void stage6_update(void);
+
+#endif // IGUARD_stages_stage6_background_anim_h
