+++ src/stages/stage6/meson.build
@@ -3,6 +3,21 @@ stage6_src = files(
 	'background_anim.c',
 	'elly.c',
 	'draw.c',
+    'extra.c',
 	'stage6.c',
 	'timeline.c',
+    'nonspells/boss_nonspell_1.c',
+    'nonspells/boss_nonspell_2.c',
+    'nonspells/boss_nonspell_3.c',
+    'nonspells/boss_nonspell_4.c',
+    'nonspells/boss_nonspell_5.c',
+    'spells/lhc.c',
+    'spells/toe.c',
+    'spells/ricci.c',
+    'spells/kepler.c',
+    'spells/newton.c',
+    'spells/broglie.c',
+    'spells/maxwell.c',
+    'spells/forgotten.c',
+    'spells/eigenstate.c'
 )
