+++ src/stages/stage6/nonspells/nonspells.h
@@ -1,0 +1,23 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
+ */
+
+#ifndef IGUARD_stages_stage6_nonspells_nonspells_h
+#define IGUARD_stages_stage6_nonspells_nonspells_h
+
+#include "taisei.h"
+
+#include "boss.h"
+
+void elly_frequency(Boss*, int);
+void elly_frequency2(Boss*, int);
+void elly_paradigm_shift(Boss*, int);
+void elly_baryonattack(Boss*, int);
+void elly_baryonattack2(Boss*, int);
+void elly_baryon_explode(Boss*, int);
+
+#endif // IGUARD_stages_stage6_nonspells_nonspells_h
