+++ src/stages/stage6/spells/maxwell.c
@@ -1,0 +1,55 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
+*/
+
+#include "taisei.h"
+
+#include "spells.h"
+#include "../elly.h"
+
+#include "common_tasks.h"
+#include "global.h"
+
+static cmplx maxwell_laser(Laser *l, float t) {
+    if(t == EVENT_BIRTH) {
+        l->unclearable = true;
+        l->shader = r_shader_get_optional("lasers/maxwell");
+        return 0;
+    }
+
+    return l->pos + l->args[0]*(t+I*creal(l->args[2])*t*0.02*sin(0.1*t+cimag(l->args[2])));
+}
+
+static void maxwell_laser_logic(Laser *l, int t) {
+    static_laser(l, t);
+    TIMER(&t);
+
+    if(l->width < 3)
+        l->width = 2.9;
+
+    FROM_TO(60, 99, 1) {
+        l->args[2] += 0.145+0.01*global.diff;
+    }
+
+    FROM_TO(00, 10000, 1) {
+        l->args[2] -= 0.1*I+0.02*I*global.diff-0.05*I*(global.diff == D_Lunatic);
+    }
+}
+
+void elly_maxwell(Boss *b, int t) {
+    TIMER(&t);
+
+    AT(250) {
+        elly_clap(b,50);
+            play_sfx("laser1");
+
+    }
+    FROM_TO(40, 159, 5) {
+        create_laser(b->pos, 200, 10000, RGBA(0, 0.2, 1, 0.0), maxwell_laser, maxwell_laser_logic, cexp(2.0*I*M_PI/24*_i)*VIEWPORT_H*0.005, 200+15.0*I, 0, 0);
+    }
+
+}
