+++ src/stages/stage6/timeline.h
@@ -1,0 +1,19 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
+ */
+
+#ifndef IGUARD_stages_stage6_timeline_h
+#define IGUARD_stages_stage6_timeline_h
+
+#include "taisei.h"
+
+#include "boss.h"
+
+void stage6_events(void);
+Boss* stage6_spawn_elly(cmplx);
+
+#endif // IGUARD_stages_stage6_timeline_h
